+++ DelvUI/Interface/PartyCooldowns/PartyCooldownsConfig.cs
@@ -312,7 +312,7 @@ namespace DelvUI.Interface.PartyCooldowns
                             CooldownsDataChangedEvent?.Invoke(this);
                         }
 
-                        if (ImGui.IsItemHovered()) { ImGui.SetTooltip("Priority determines which cooldows show first on the list."); }
+                        ImGuiHelper.SetTooltip("Priority determines which cooldows show first on the list.");
                     }
 
                     // column
@@ -327,7 +327,7 @@ namespace DelvUI.Interface.PartyCooldowns
                             CooldownsDataChangedEvent?.Invoke(this);
                         }
 
-                        if (ImGui.IsItemHovered()) { ImGui.SetTooltip("Allows to separate cooldowns in different columns."); }
+                        ImGuiHelper.SetTooltip("Allows to separate cooldowns in different columns.");
                     }
 
                     ImGui.PopID();